+++ src/main/java/com/hapfu/course/data/model/one/generators/RandomGenerator.java
@@ -1,7 +1,9 @@
 package com.hapfu.course.data.model.one.generators;
 
+import java.math.BigDecimal;
+
 public interface RandomGenerator {
 
-    double generate();
+    BigDecimal generate();
 
 }
